pc11 -> pc61 upgrade work
This includes:

rewriting the exist pc11->pc61 to be more understandable (and may
be work better).

Add stats to this process and provide show/spotstats to enquire
how it's all going.

Dump the routing table to /spider/local_data (in json) every 10
minutes and on exit. Read the routing table back in on start up
so that all that IP address info isn't lost over software updates
or other routine restarts.